Initial version of json schema for TCG's TPM TSS JSON Policy Language
This repository aims to provide a suitable json schema to validate policies
which can be used with the tpm2-tss fapi layer and tools.

This commit adds a first draft to show the general structure of a json schema
which matches some basic policies.

The work is sponsored by Infineon Technologies AG, but is considered an open project
and will later move into the tpm2-software.github.io ecosystem.
Patches welcome.

Signed-off-by: Peter Huewe <peter.huewe@infineon.com>